bigger cleanup with findbugs
git-svn-id: http://svn.l2jfree.com/l2jfree@4150 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/model/actor/instance/L2PetInstance.java
@@ -459,9 +459,9 @@ public class L2PetInstance extends L2Summon
            return;
        }
 
-       int weight=ItemTable.getInstance().getTemplate(target.getItemId()).getWeight() * target.getCount();
+       long weight=ItemTable.getInstance().getTemplate(target.getItemId()).getWeight() * target.getCount();
        
-       if (weight > Integer.MAX_VALUE || weight < 0 || !getInventory().validateWeight(weight))
+       if (weight > Integer.MAX_VALUE || weight < 0 || !getInventory().validateWeight((int)weight))
        {
            sendPacket(new SystemMessage(SystemMessageId.YOUR_PET_CANNOT_CARRY_ANY_MORE_ITEMS));
            return;
